==> BD_nueva/3_TransformacionesFinales.Rmd <==
---
title: "Transformaciones finales de la base de datos"
author: "Equipo de TAE"
date: "Diciembre de 2021"
output:
  html_document:
    code_folding: show
    toc: true
    theme: paper
    df_print: paged
    number_sections: true
---

<style>
body {
text-align: justify}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

```{r, message = FALSE}
library("DT")
library("tidyverse")
library("lubridate")
```

```{r}
# Lectura de la base de datos
#load("accidentesMDE.RData")
accidentes <- read.csv("accidentesConBarrio3_1.csv", header = TRUE, sep = ",",
                        dec = ".", encoding = "UTF-8")
```

```{r}
# Conversión de columnas de carácter a factor
accidentes[sapply(accidentes, is.character)] <- lapply(accidentes[sapply(accidentes, is.character)], 
                                       as.factor)
```

# Introducción

Una vez se ha hecho la depuración de la base de datos y se han realizado las imputaciones, fundementalmente en los barrios, se puede proceder con las transformaciones finales de la tabla de datos de cara al estudio de los diferentes modelos predictivos que van a ser abordados para poder proyectar la accidentalidad en la ciudad de Medellín durante los años 2020 y 2021.

## Simplificación de columnas

Para comenzar, se van a eliminar las columnas relacionadas con datos administrativos, como lo son el CBML, la dirección y la dirección encasillada, el número de radicado, el número de la comuna, el expediente y la fecha de accidente secundaria. La justificación de la eliminación de esta última columna se basa en que, al revisar las observaciones, se evidencia que la hora asociada está seis o cinco horas más tarde y generalmente en horarios nocturnos en los que la circulación vehicular es muy reducida en la ciudad, lo que lleva a pensar que el huso horario de esta variable está en la hora de Londres (*Greenwich Meridian Time*, ***GMT***), toda vez que la primera fecha y hora reportadas presentan un huso horario adecuado y coherente con los horarios y hábitos de la ciudadanía en Medellín, cuyos picos de circulación se dan entre las [8:00 am y las 6.00 pm](http://www.scielo.org.co/scielo.php?script=sci_arttext&pid=S0012-73532011000100011#:~:text=Con%20esta%20información%20de%20los,y%20las%2018%3A00%20horas.)

```{r}
# Eliminación de datos administrativos
accidentes <- accidentes %>%
  select(-c("DIRECCION", "EXPEDIENTE", "NRO_RADICADO", "NUMCOMUNA",
            "DIRECCION.ENCASILLADA", "FECHA_ACCIDENTES", "HORAS", "LOCATION"))
```

```{r}
# Reubicación
accidentes <- accidentes %>% 
  relocate(FECHA_ACCIDENTE, .before = CLASE_ACCIDENTE) %>% 
  relocate(MES, .before = CLASE_ACCIDENTE) %>% 
  relocate(FECHA_ACCIDENTE, .before = AÑO) %>% 
  relocate(MES, .before = CLASE_ACCIDENTE) %>% 
  relocate(FECHA, .before = CLASE_ACCIDENTE) %>% 
  relocate(HORA, .before = CLASE_ACCIDENTE)
```


## Comuna

Como se vio al finalizar el primer proceso de depuración, habían varias decenas de barrios en la base de datos desconocidos, y para no dejar sus campos vacíos ni descartarlos para el modelo predictivo se realizó un proceso de imputación usando el método de los vecinos más cercanos, lo cual puede generar algunas diferencias entre el barrio y su comuna en tanto en algunas observaciones el barrio era desconocido pero no su columna, por lo que se debe proceder a la modifación de las columnas para que estos valores coincidan.

```{r}
# Vectores con los barrios para cada comuna y corregimiento de Medellín

## Comuna No. 01 - El Popular

com01 <- c("Aldea Pablo VI", "Carpinelo", "El Compromiso", "Granizal",
           "La Avanzada", "La Esperanza No. 2", "Moscú No. 2", "Popular",
           "Santo Domingo Savio No. 1", "Santo Domingo Savio No. 2",
           "San Pablo", "Villa Guadalupe")

## Comuna No. 02 - Santa Cruz

com02 <- c("Santa Cruz", "La Isla", "Playón de los Comuneros", "Pablo VI",
           "La Frontera", "La Francia", "Andalucía", "Villa del Socorro",
           "Villa Niza", "Moscú No. 1", "La Rosa")

## Comuna No. 03 - Manrique

com03 <- c("La Salle", "Las Granjas", "Campo Valdés No. 2", "Santa Inés",
           "El Raizal", "El Pomar", "Manrique Central No. 2", "Manrique Oriental",
           "Versalles No. 1", "Versalles No. 2", "La Cruz", "La Honda",
           "Oriente", "María Cano Carambolas", "San José de la Cima No. 1",
           "San José de la Cima No. 2")

## Comuna No. 04 - Aranjuez

com04 <- c("Aranjuez", "Berlín", "San Isidro", "Palermo", "Bermejal - Los Álamos",
           "Moravia", "Sevilla", "San Pedro", "Manrique Central No. 1",
           "Campo Valdés No. 1", "Las Esmeraldas", "La Piñuela", "Brasilia",
           "Miranda", "Jardín Botánico", "Universidad de Antioquia",
           "Parque Norte")

## Comuna No. 05 - Castilla

com05 <- c("Castilla", "Toscana", "Héctor Abad Gómez", "Las Brisas", "Florencia",
           "Tejelo", "Boyacá", "Belalcázar", "Girardot", "Tricentenario",
           "Francisco Antonio Zea", "Alfonso López", "Caribe", "El Progreso",
           "Plaza de Ferias", "Cementerio Universal", "Oleoducto",
           "Terminal de Transporte")

## Comuna No. 06 - Doce de Octubre

com06 <- c("Doce de Octubre No. 1", "Doce de Octubre No. 2", "Santander",
           "Pedregal", "La Esperanza", "San Martín de Porres", "Kennedy",
           "Picacho", "Picachito", "Mirador del Doce", "El Progreso No. 2",
           "El Triunfo")

## Comuna No. 07 - Robledo

com07 <- c("Robledo", "Barrio Cerro El Volador", "San Germán", "Facultad de Minas U. Nacional",
           "La Pilarica", "Bosques de San Pablo", "Altamira", "Córdoba",
           "López de Mesa", "El Diamante", "Aures No. 1", "Aures No. 2",
           "Bello Horizonte", "Villa Flora", "Palenque", "Cucaracho",
           "Fuente Clara", "Santa Margarita", "Olaya Herrera", "Pajarito",
           "Monteclaro", "Nueva Villa de la Iguaná", "La Cuchilla",
           "La Aurora", "La Campiña", "Universidad Nacional",
           "Facultad Veterinaria y Zootecnia U.de.A.")

## Comuna No. 08 - Villa Hermosa

com08 <- c("Villa Hermosa", "La Mansión", "Llanaditas", "La Ladera",
           "Golondrinas", "Batallón Girardot", "San Miguel", "Los Mangos",
           "Sucre", "El Pinal", "Trece de Noviembre", "La Libertad",
           "Villatina", "San Antonio", "Las Estancias", "Villa Turbay",
           "La Sierra", "Villa Lilliam", "Esfuerzos de Paz No. 1",
           "Esfuerzos de Paz No. 2", "Enciso")

## Comuna No. 09 - Buenos Aires

com09 <- c("Buenos Aires", "Barrio Caicedo", "Juan Pablo II", "Ocho de Marzo",
           "Barrio de Jesús", "Bomboná No. 2", "Los Cerros El Vergel",
           "Alejandro Echavarría", "Miraflores", "Cataluña", "La Milagrosa",
           "Gerona", "El Salvador", "Loreto", "Asomadera No. 1", "Asomadera No. 2",
           "Asomadera No. 3", "Quinta Linda")

## Comuna No. 10 - La Candelaria

com10 <- c("La Candelaria", "Prado", "Jesús Nazareno", "El Chagualo",
           "Estación Villa", "San Benito", "Guayaquil", "Corazón de Jesús",
           "Calle Nueva", "Perpetuo Socorro", "Barrio Colón", "Las Palmas",
           "Bomboná No. 1", "Boston", "Los Ángeles", "Villa Nueva",
           "San Diego", "Centro Administrativo", "La Alpujarra",
           "Hospital San Vicente de Paúl")

## Comuna No. 11 - Laureles - Estadios

com11 <- c("Los Conquistadores", "Laureles", "Carlos E. Restrepo",
           "Suramericana", "Naranjal", "San Joaquín", "Bolivariana",
           "Las Acacias", "La Castellana", "Lorena", "El Velódromo",
           "El Estadio", "Los Colores", "Cuarta Brigada", "Florida Nueva",
           "U.P.B.", "Unidad Deportiva Atanasio Girardot")

## Comuna No. 12 - La América

com12 <- c("La América", "Ferrini", "Calasanz", "Los Pinos", "La Floresta",
           "Santa Lucía", "El Danubio", "Campo Alegre", "Santa Mónica",
           "Barrio Cristóbal", "Simón Bolívar", "Santa Teresita",
           "Calasanz Parte Alta")

## Comuna No. 13 - San Javier

com13 <- c("San Javier No. 1", "San Javier No. 2", "El Pesebre", "Blanquizal",
           "Santa Rosa de Lima", "Los Alcázares", "Metropolitano", "La Pradera",
           "Juan XXIII La Quiebra", "La Divisa", "Veinte de Julio", "Belencito", "Betania",
           "El Corazón", "Las Independencias", "Nuevos Conquistadores",
           "El Salado", "Eduardo Santos", "Peñitas", "Antonio Nariño",
           "El Socorro", "Calasania", "La Asomadera")

## Comuna No. 14 - El Poblado

com14 <- c("Castropol", "Barrio Colombia", "Villa Carlota", "Lalinde",
           "Manila", "Las Lomas No. 1", "Las Lomas No. 2", "Altos del Poblado",
           "El Tesoro", "Los Naranjos", "Los Balsos No. 1",
           "Los Balsos No. 2", "San Lucas", "El Diamante No. 2", "El Castillo",
           "Alejandría", "La Florida", "El Poblado", "Astorga", "Patio Bonito",
           "La Aguacatala", "Santa María de los Ángeles")

## Comuna No. 15 - Guayabal

com15 <- c("Tenche", "Trinidad", "Santa Fe", "Campo Amor", "Cristo Rey",
           "Guayabal", "La Colina", "El Rodeo", "Parque Juan Pablo II")

## Comuna No. 16 - Guayabal 

com16 <- c("Belén", "Rodeo Alto", "Aguas Frías", "Cerro Nutibara", "Fátima",
           "Rosales", "Granada", "San Bernardo", "Las Playas", "Diego Echavarría",
           "La Mota", "El Rincón", "La Hondonada", "Loma de los Bernal",
           "La Gloria", "Altavista", "La Palma", "Zafra", "Los Alpes",
           "Las Violetas", "Las Mercedes", "Nueva Villa de Aburrá", "Miravalle",
           "El Nogal - Los Almendros")

## Corregimiento No. 50 - Palmitas

com50 <- c("La Aldea", "Suburbano Palmitas", "La Frisola", "Volcana Guayabal",
           "Palmitas Sector Central", "La Suiza", "Potrera Miserenga",
           "Suburbano La Aldea", "La Urquita", "Suburbano Potreta Miserenga",
           "Suburbano Urquita")

## Corregimiento No. 60 - San Cristóbal

com60 <- c("Boquerón", "Suburbano El Llano", "Cabecera Urbana San Cristóbal",
           "Suburbano La Palma - El Patio", "La Loma", "Área de expansión Altos de Calasanz",
           "Travesías", "Pedregal Alto", "El Llano", "Pajarito", "La Loma Oriental", "Naranjal",
           "El Uvito", "Las Playas", "Suburbano Travesías", "Suburbano La Loma", "Pedregal Bajo",
           "La Palma - El Patio", "Área de expansión Pajarito", "Suburbano La Cuchilla", "El Carmelo",
           "Yolombo", "San José de la Montaña", "El Picacho", "La Cuchilla", "Suburbano Pedregal Alto",
           "AUC1", "AUC2")

## Corregimiento No. 70 - Altavista

com70 <- c("Altavista Sector Central", "Área de expansión Altavista", "Aguas Frías",
           "Área de expansión El Noral", "San Pablo", "El Corazón - El Morro",
           "Área de expansión Belén Rincón", "San José del Manzanillo", "Suburbano Aguas Frías",
           "La Esperanza", "Buga Patio Bonito", "Suburbano Altavista")

## Corregimiento No. 80 - San Antonio de Prado

com80 <-  c("La Florida", "Área de expansión San Antonio de Prado",
            "Suburbano Potrerito", "La Verde", "El Astillero", "El Salado",
            "Potrerito", "Yarumalito", "Cabecera San Antonio de Prado",
            "La Oculta", "El Vergel")

## Corregimiento No. 90 - Santa Elena

com90 <- c("Piedra Gorda", "Suburbano Matasano No. 1", "Suburbano Piedra Gorda",
           "Piedras Blancas Represa", "Santa Elena Sector Central", "Barro Blanco",
           "Suburbano El Placer", "El Placer", "Suburbano Santa Elena Central",
           "Suburbano El Cerro", "El Plan", "El Cerro", "Suburbano Barro Blanco",
           "Media Luna", "Suburbano El Llano", "Suburbano Chacaltaya", "Suburbano Matasano No. 2",
           "Piedras Blancas", "Suburbano El Plan", "Suburbano El Tesoro",
           "Suburbano Matasano No. 3", "El Llano SE", "Suburbano Mirador del Poblado")

## Medellín

MDE <- c(com01, com02, com03, com04, com05, com06, com07, com08, com09, com10,
         com11, com12, com13, com14, com15, com16, com50, com60, com70, com80,
         com90)
```

```{r}
accidentes <- accidentes %>% 
  mutate(COMUNASMOD = ifelse(BARRIO %in% com01, "Popular",
                      ifelse(BARRIO %in% com02, "Santa Cruz",
                      ifelse(BARRIO %in% com03, "Santa Manrique",
                      ifelse(BARRIO %in% com04, "Aranjuez",
                      ifelse(BARRIO %in% com05, "Castilla",
                      ifelse(BARRIO %in% com06, "Doce de Octubre",
                      ifelse(BARRIO %in% com07, "Robledo",
                      ifelse(BARRIO %in% com08, "Villa Hermosa",
                      ifelse(BARRIO %in% com09, "Buenos Aires",
                      ifelse(BARRIO %in% com10, "La Candelaria",
                      ifelse(BARRIO %in% com11, "Laureles - Estadio",
                      ifelse(BARRIO %in% com12, "La América",
                      ifelse(BARRIO %in% com13, "San Javier",
                      ifelse(BARRIO %in% com14, "El Poblado",
                      ifelse(BARRIO %in% com15, "Guayabal",
                      ifelse(BARRIO %in% com16, "Belén",
                      ifelse(BARRIO %in% com50, "San Sebastián de Palmitas",
                      ifelse(BARRIO %in% com60, "San Cristóbal",
                      ifelse(BARRIO %in% com70, "Altavista",
                      ifelse(BARRIO %in% com80, "San Antonio de Prado",
                      ifelse(BARRIO %in% com90, "Santa Elena",
                             "NA"))))))))))))))))))))))
```

```{r}
problemas <- accidentes %>% 
  filter(COMUNASMOD == NA | COMUNASMOD == "" | COMUNASMOD == "x") %>% 
  group_by(BARRIO) %>% 
  summarise(TOTAL = n())
```



## Fecha y fechas especial.

Ahora bien, la variable que relaciona la fecha en la que el incidente vial ocurrió es de suma importancia ya que lo que se quiere predecir es precisamente la cantidad y tipo de accidentes que ocurren en la ciudad de Medellín en el 2020 y el 2021 a partir de la información anterior, y para poder facilitar esto se va a a extraer información importante de las fechas registradas, como el día, el mes, el año, el número de la semana y el día de la semana en la que se dio el accidente. Adicionalmente, se van a crear varias variables binarias que van a identificar qué accidentes se dieron en fechas que cumplen las siguientes características.

- **Festivo.** Los días festivos son de importancia ya que en ellos se dan operaciones de retorno de miles de ciudadanos que previamente salieron de la ciudad hacia otros municipios de la región y el país, lo cual podría explicar varios accidentes registrados en la base de datos. También se agregan a nochebuena (24 de diciembre de cada año) y nochevieja (31 de diciembre de cada año).
- **Feria de flores.** La feria de flores es el evento más icónico de la ciudad de Medellín y uno de los más reconocidos de Colombia, lo cual podría estimular el uso de vehículos para asistir a diferentes eventos de esta feria, así como el tránsito de más personas que llegan de otros lugares del país o del mundo para participar en las actividades de la feria de flores.
- **Final de fútbol.** La ciudad de Medellín tiene dos equipos de fútbol: el Atlético Nacional y el Deportivo Independiente Medellín, los cuales han participado en varias finales de la categoría A del torneo de fútbol nacional, por lo que es conveniente considerarla ya que por la movilización de miles de personas que desean reunirse con familiares y amigos para disfrutar de la cita deportiva juntos podría explicar una alta cantidad de accidentes.
- **Quincena.** Se incluyen las quincenas tradicionales, que son el quince y el treinta de cada mes, así como el primero, ya que algunas empresas tienen esquema primero - quince, y también a los días cinco y al veinte de cada mes, puesto que otras empresas y sobre todo la administración pública paga a sus empleados dichos días.

Adicionalmente, para facilitar la manipulación de las fechas, se generan a partir de la fecha reportada en el IPAT las siguientes columnas:

- **Día modificado.** Componente del día de la fecha en la que se registró en el IPAD la observación de un accidente de tránsito.
- **Mes modificado.** Componente del mes de la fecha en la que se registró en el IPAD la observación de un accidente de tránsito.
- **Año modificado.** Componente del año de la fecha en la que se registró en el IPAD la observación de un accidente de tránsito.
- **Semana modificada uno.** Número de la semana en el año en la que se dio un accidente de tránsito observado, teniendo en cuenta que la primera semana va del primero a siete de enero, la segunda va del ocho al catorce de enero y así sucesivamente sin considerar el día de la semana asociada a cada día.
- **Semana modificada dos.** Número de la semana del año en la que se dio un accidente de tránsito observado considerando el día en el que comienza el año. Así, si un año comienza el día sabado, entonces el primero y el segundo de enero son la primera semana, pero el tres de enero, siendo lunes, registra una nueva semana, que en este caso es la segunda, y así sucesivamente.
- **Día semana.** Nombre del día de la semana en la que se dio un accidente de tránsito observado.
- **Mes semana.** Mes en el que se dio un accidente de tránsito según la fecha formalmente registrada en el IPAT.
- **Festivo.** Identificación de días asociados a festivos.
- **Flores.** Identificación de días asociados a la celebración de la Feria de las Flores en la ciudad.
- **Fútbol.** Identificación de días en los que algún equipo local de fútbol participó en una final de fútbol de la categoría primera del país.
- **Quincena.** Identificación de días en los que las empresas y las instituciones públicas y privadas pagan a sus empleados formalos sus quincenas.

```{r}
accidentes <- accidentes %>% 
  mutate(FECHA = as.Date(FECHA_ACCIDENTE)) %>% 
  mutate(DIAX = day(FECHA)) %>% 
  mutate(MESX = month(FECHA)) %>% 
  mutate(AÑOX = year(FECHA)) %>% 
  mutate(SEMANAX = week(FECHA)) %>% # 1-7, 8-14,...
  mutate(SEMANAXX = strftime(FECHA, format = "%V")) %>% # Semana del año sin importar inicio
  mutate(DIA_SEMANA = weekdays(FECHA)) %>% 
  mutate(MES_SEMANA = months(FECHA)) 
```

```{r}
# Festivos
fest14 <- c("2014-01-01", "2014-01-06", "2014-03-24",
                "2014-04-17", "2014-04-18", "2014-05-01",
	            	"2014-06-02", "2014-06-23", "2014-06-30",
            		"2014-07-20", "2014-08-07", "2014-08-18",
            		"2014-10-13", "2014-11-03", "2014-11-17",
            		"2014-12-08", "2014-12-25")

fest15 <- c("2015-01-01", "2015-01-12", "2015-03-23",
            "2015-04-02", "2015-04-03", "2015-05-01",
        		"2015-06-08", "2015-06-15", "2015-06-29",
        		"2015-07-20", "2015-08-07", "2015-08-17",
        		"2015-10-12", "2015-11-02", "2015-11-16",
        		"2015-12-08", "2015-12-25", "2015-05-18")

fest16 <- c("2016-01-01", "2016-01-11", "2016-03-21",
            "2016-03-24", "2016-03-25", "2016-05-01",
        		"2016-05-30", "2016-06-06", "2016-07-04",
        		"2016-07-20", "2016-08-07", "2016-08-15",
        		"2016-10-17", "2016-11-07", "2016-11-14",
        		"2016-12-08", "2016-12-25", "2016-05-09")

fest17 <- c("2017-01-01", "2017-01-09", "2017-03-20",
                "2017-04-13", "2017-04-14", "2017-05-01",
		"2017-05-29", "2017-06-19", "2017-07-03",
		"2017-07-20", "2017-08-07", "2017-08-21",
		"2017-10-16", "2017-11-06", "2017-11-13",
		"2017-12-08", "2017-12-25", "2017-06-26")

fest18 <- c("2018-01-01", "2018-01-08", "2018-03-19",
            "2018-03-19", "2018-03-30", "2018-05-01",
        		"2018-05-14", "2018-06-04", "2018-07-02",
        		"2018-07-20", "2018-08-07", "2018-08-20",
        		"2018-10-15", "2018-11-05", "2018-11-12",
        		"2018-12-08", "2018-12-25", "2018-06-11")

fest19 <- c("2019-01-01", "2019-01-07", "2019-03-25",
            "2019-04-19", "2019-04-18", "2019-05-01",
        		"2019-06-03", "2019-06-24", "2019-07-01",
        		"2019-07-20", "2019-08-07", "2019-08-19",
        		"2019-10-14", "2019-11-04", "2019-11-11",
        		"2019-12-08", "2019-12-25")

fest20 <- c("2020-01-01", "2020-01-06", "2020-03-23",
            "2020-04-10", "2020-04-09", "2020-05-01",
        		"2020-06-15", "2020-06-22", "2020-06-29",
        		"2020-07-20", "2020-08-07", "2020-08-17",
        		"2020-10-12", "2020-11-02", "2020-11-16",
        		"2020-12-08", "2020-12-25", "2020-05-25")

festivos <- c(fest14,fest15,fest15,fest16,fest17,fest18,fest19,fest20)

festivos <- as.Date(festivos)

accidentes <- accidentes %>% 
  mutate(FESTIVO = ifelse((FECHA %in% festivos), 1, 0))
```

```{r}
# Feria de Flores

#2014
ff14 <- seq(as.Date("2014-08-01"), by = "day", length.out = 10)

#2015
ff15 <- seq(as.Date("2015-07-31"), by = "day", length.out = 10)

#2016
ff16 <- seq(as.Date("2016-07-29"), by = "day", length.out = 10)

#2017
ff17 <- seq(as.Date("2017-07-28"), by = "day", length.out = 10)

#2018
ff18 <- seq(as.Date("2018-08-03"), by = "day", length.out = 10)

#2019
ff19 <- seq(as.Date("2015-08-02"), by = "day", length.out = 10)

#2020
ff20 <- seq(as.Date("2020-11-01"), by = "day", length.out = 8)

#2021
ff21 <- seq(as.Date("2021-08-12"), by = "day", length.out = 8)

ff <- c(ff14, ff15, ff16, ff17, ff18, ff19, ff20, ff21)

accidentes <- accidentes %>% 
  mutate(FLORES = ifelse((FECHA %in% ff), 1, 0))
```

```{r}
# Fútbol
futbol <- c("2014-12-17", "2014-12-17",
            "2015-06-07", "2015-06-03",
            "2015-12-16", "2015-12-20",
            "2016-06-15", "2016-06-19",
            "2017-06-14", "2017-06-18",
            "2018-06-06", "2018-06-09",
            "2018-12-13", "2018-12-17",
            "2016-01-23", "2016-01-27",
            "2017-01-21", "2017-01-29",
            "2018-01-31", "2019-02-07")

futbol <- as.Date(futbol)

accidentes <- accidentes %>% 
  mutate(FUTBOL = ifelse((FECHA %in% futbol), 1, 0))
```

```{r}
# Quincena
quincena <- c("5", "15", "20", "30", "1")

accidentes <- accidentes %>% 
  mutate(QUINCENA = ifelse((DIAX %in% quincena), 1, 0))
```

```{r, include = FALSE, warning = FALSE, echo = FALSE}
# Identificación de viernes y sábado previo a un día festivo que es lunes

library("zoo")
vs_fest <- as.Date(character())

for (i in festivos) {
    i = as.Date(i)
    if(weekdays(i) == "lunes") {
      vs_fest <- c(vs_fest, (i -1)) # Sábado
      vs_fest <- c(vs_fest, (i -2)) # Domingo
      vs_fest <- c(vs_fest, (i -3)) # Viernes
    }
}
```


```{r}
accidentes <- accidentes %>% 
  mutate(PREFESTIVO = ifelse((FECHA %in% vs_fest), 1, 0))
```


## Tipo de accidente por fecha

Finalmente, para facilitar el estudio de los diferentes tipos de accidente y del diseño de vías, se generan unas bases de datos que agrupan el tipo de accidente sucedido sumando las observaciones que tienen en común aspectos como la fecha, el diseño de la vía o el barrio.

```{r, message = FALSE}
accFechaAcc <- accidentes %>% 
  group_by(FECHA, CLASE_ACCIDENTE) %>% 
  summarise(ACCIDENTES = n()) %>% 
  mutate(DIAX = day(FECHA)) %>% 
  mutate(MESX = month(FECHA)) %>% 
  mutate(AÑOX = year(FECHA)) %>% 
  mutate(SEMANAX = week(FECHA)) %>% # 1-7, 8-14,...
  mutate(SEMANAXX = strftime(FECHA, format = "%V")) %>% # Semana del año sin importar inicio
  mutate(DIA_SEMANA = weekdays(FECHA)) %>% 
  mutate(MES_SEMANA = months(FECHA)) %>% 
  mutate(FESTIVO = ifelse((FECHA %in% festivos), 1, 0)) %>% 
  mutate(FLORES = ifelse((FECHA %in% ff), 1, 0)) %>% 
  mutate(FUTBOL = ifelse((FECHA %in% futbol), 1, 0)) %>% 
  mutate(QUINCENA = ifelse((DIAX %in% quincena), 1, 0)) %>% 
  mutate(PREFESTIVO = ifelse((DIAX %in% vs_fest), 1, 0))
```

```{r, message = FALSE}
accFechaDis <- accidentes %>%  
  group_by(FECHA, DISEÑO) %>% 
  summarise(ACCIDENTES = n()) %>% 
  mutate(DIAX = day(FECHA)) %>% 
  mutate(MESX = month(FECHA)) %>% 
  mutate(AÑOX = year(FECHA)) %>% 
  mutate(SEMANAX = week(FECHA)) %>% # 1-7, 8-14,...
  mutate(SEMANAXX = strftime(FECHA, format = "%V")) %>% # Semana del año sin importar inicio
  mutate(DIA_SEMANA = weekdays(FECHA)) %>% 
  mutate(MES_SEMANA = months(FECHA)) %>% 
  mutate(FESTIVO = ifelse((FECHA %in% festivos), 1, 0)) %>% 
  mutate(FLORES = ifelse((FECHA %in% ff), 1, 0)) %>% 
  mutate(FUTBOL = ifelse((FECHA %in% futbol), 1, 0)) %>% 
  mutate(QUINCENA = ifelse((DIAX %in% quincena), 1, 0)) %>% 
  mutate(PREFESTIVO = ifelse((DIAX %in% vs_fest), 1, 0))
```
```{r, message = FALSE}
accFechaDisAcc <- accidentes %>%  
  group_by(FECHA, DISEÑO, CLASE_ACCIDENTE) %>% 
  summarise(ACCIDENTES = n()) %>% 
  mutate(DIAX = day(FECHA)) %>% 
  mutate(MESX = month(FECHA)) %>% 
  mutate(AÑOX = year(FECHA)) %>% 
  mutate(SEMANAX = week(FECHA)) %>% # 1-7, 8-14,...
  mutate(SEMANAXX = strftime(FECHA, format = "%V")) %>% # Semana del año sin importar inicio
  mutate(DIA_SEMANA = weekdays(FECHA)) %>% 
  mutate(MES_SEMANA = months(FECHA)) %>% 
  mutate(FESTIVO = ifelse((FECHA %in% festivos), 1, 0)) %>% 
  mutate(FLORES = ifelse((FECHA %in% ff), 1, 0)) %>% 
  mutate(FUTBOL = ifelse((FECHA %in% futbol), 1, 0)) %>% 
  mutate(QUINCENA = ifelse((DIAX %in% quincena), 1, 0)) %>% 
  mutate(PREFESTIVO = ifelse((DIAX %in% vs_fest), 1, 0))
```
```{r}
save(accidentes, file = "accidentesMDE2.RData")
save(accFechaDis, file = "accFechaDiseno.RData")
save(accFechaAcc, file = "accFechaTipo.RData")
save(accFechaDisAcc, file = "accFechaDisenoTipo.RData")
write.csv(accidentes, file = "accidentesMDE2.csv")
```
