==> BD_nueva/4_Descripcion.Rmd <==
---
title: "Descripción de la accidentalidad en Medellín"
author: "Simón Cuartas Rendón"
date: "Noviembre de 2021"
output:
  html_document:
    code_folding: show
    toc: true
    theme: paper
    df_print: paged
    number_sections: true
---

<style>
body {
text-align: justify}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE)
```

# Breve introducción

En este documento se abordará una breve descripción relaciondada con la base de datos sobre la accidentalidad vehicular en la ciudad de Medellín entre los años 2014 y 2020, haciéndo énfasis en las variables más relevantes de esta base de datos como lo son la fecha de ocurrencia, las comunas y los barrios en los que se dan los incidentes y el tipo de accidentalidad sucedida, entre otros. Esto con el propósito de tener un acercamiento inicial a esta base de datos como parte del primer proyecto del curso de ***Técnicas de Aprendizaje Estadístico*** del semestre 2021-2 de la Universidad Nacional de Colombia, sede Medellín.

```{r Lectura de paquetes, include = FALSE}
library("tidyverse")
library("kableExtra")
library("plotly")
library("summarytools")
library("DT")
library("iClick")
library("lubridate")
library("formattable")
library("nortest")
```

```{r Lectura de la base de datos}
load("accidentesMDE2.RData")
```

```{r, include = FALSE}
glimpse(accidentes)
```

| Característica         | Formato  | Observación uno     |
|------------------------|----------|---------------------|
| Fecha del accidente    | Factor   | 2014-12-13 17:00:00 |
| Año                    | Doble    | 2014                |
| Mes                    | Factor   | Diciembre           |
| Fecha                  | Fecha    | 2014-12-13          |
| Hora                   | Factor   | 17:00:00            |
| Clase de accidente     | Factor   | Atropello           |
| Diseño                 | Factor   | Tramo de vía        |
| Gravedad del accidente | Factor   | Con heridos         |
| Barrio                 | Factor   | Los Ángeles         |
| Comuna                 | Factor   | NA                  |
| Coordenada x, Magna    | Doble    | 838649.             |
| Coordenada Y, Magna    | Doble    | 1180876             |
| Longitud               | Doble    | -75.53534           |
| Latitud                | Doble    | 6.229826            |
| Comunas modificado     | Carácter | La Candelaria       |
| Día modificado         | Entero   | 13                  |
| Mes modificado         | Doble    | 12                  |
| Año modificado         | Doble    | 2014                |
| Semana modificado      | Doble    | 50                  |
| Semana modificado II   | Carácter | 50                  |
| Día semana             | Carácter | Sábaddo             |
| Mes semana             | Carácter | Diciembre           |
| Festivo                | Doble    | 0                   |
| Flores                 | Doble    | 0                   |
| Fútbol                 | Doble    | 0                   |
| Quincena               | Doble    | 0                   |
| Puente                 | Doble    | 0                   |

# Descripción general de la base de datos

La base de datos de accidentalidad de Medellín recoge todos los accidentes que sucedieron en la ciudad de Medellín, Antioquia (Colombia) entre el lunes 14 de julio de 2014 y el lunes 31 de agosto de 2020, recogiendo algunos aspectos de cada uno de estos incidentes viales como la clase de accidente observado, la dirección en la que ocurrió, el código del expediente y su número de radicado, la fecha y hora y la ubicación en coordenadas geográficas. En general, las variables a considerar por esta base de datos son las siguientes:

- **Fecha del accidente.** Fecha en la que sucedió el incidente de tránsito según el Informe Policial de Accidentes de Tránsito, IPAT, en el formato YYYY-MM-DD HH:MM:SS.
- **Año.** Año en el que se dio el accidente según el reporte oficial del IPAT.
- **Mes.** Mes en el que se dio el accidente según el reporte oficial del IPAT descrito en palabras (ej. *enero*)
- **Fecha.** Extracción del día, mes y año en el que se dio el incidente de tránsito según el IPAT extraído de la ***fecha del accidente*** en formato YYYY-MM-DD.
- **Hora.** Extracción de la hora y el minuto en el que se dio el incidente de tránsito según el IPAT extraído de la ***fecha del accidente*** en formato HH:MM:SS.
- **Clase de accidente.** Clasificación del IPAT sobre la clase de accidente de transito: choque, atropello, volcamiento, caida de ocupante, incendio, u otro (que no corresponde a las anteriores 5 clasificaciones, p. ej: sumersión).
- **Diseño.** Sitio de la vía donde ocurrió el accidente: ciclorruta, glorieta, intersección, lote o predio, paso a nivel, paso elevado, paso inferior, pontón, puente, tramo de vía, túnel, vía peatonal.
- **Gravedad del accidente.** Clasificación del IPAT sobre la gravedad del accidente, que corresponde al resultado más grave presentado en el accidente. Daños materiales: solo daños; accidente con heridos: hubo heridos como resultado del incidente; accidente con muertos: al menos una persona falleció como consecuencia del incidente de tránsito. No indica cantidad.
- **Barrio.** Nombre del barrio o zona en la que sucedió el accidente de tránsito.
- **Comuna** Comuna o corregimiento de Medellín a la que pertenece el barrio o zona en la que se ocurrió el accidente de tránsito observado conforme al registro en el IPAT.
- **X, magna.** Coordenadas horizontales del punto en el que sucedió el accidente de tránsito registrado con origen en el sistema de coordenadas geográficas Magna - Medellín.
- **Y, magna.** Coordenadas verticales del punto en el que sucedió el accidente de tránsito registrado con origen en el sistema de coordenadas geográficas Magna - Medellín.
- **Longitud.** Coomponente de la longitud de las coordenadas geográficas del punto en el que se dio el accidente de tránsito registrado.
- **Latitud.** Coomponente de la latitud de las coordenadas geográficas del punto en el que se dio el accidente de tránsito registrado.
- **Comuna modificada.** Verificación de la comuna a la que pertenece un barrio o zona según los mapas oficiales de la Alcaldía de Medellín.
- **Día modificado.** Componente del día de la fecha en la que se registró en el IPAD la observación de un accidente de tránsito.
- **Mes modificado.** Componente del mes de la fecha en la que se registró en el IPAD la observación de un accidente de tránsito.
- **Año modificado.** Componente del año de la fecha en la que se registró en el IPAD la observación de un accidente de tránsito.
- **Semana modificada uno.** Número de la semana en el año en la que se dio un accidente de tránsito observado, teniendo en cuenta que la primera semana va del primero a siete de enero, la segunda va del ocho al catorce de enero y así sucesivamente sin considerar el día de la semana asociada a cada día.
- **Semana modificada dos.** Número de la semana del año en la que se dio un accidente de tránsito observado considerando el día en el que comienza el año. Así, si un año comienza el día sabado, entonces el primero y el segundo de enero son la primera semana, pero el tres de enero, siendo lunes, registra una nueva semana, que en este caso es la segunda, y así sucesivamente.
- **Día semana.** Nombre del día de la semana en la que se dio un accidente de tránsito observado.
- **Mes semana.** Mes en el que se dio un accidente de tránsito según la fecha formalmente registrada en el IPAT.
- **Festivo.** Identificación de días asociados a festivos.
- **Flores.** Identificación de días asociados a la celebración de la Feria de las Flores en la ciudad.
- **Fútbol.** Identificación de días en los que algún equipo local de fútbol participó en una final de fútbol de la categoría primera del país.
- **Quincena.** Identificación de días en los que las empresas y las instituciones públicas y privadas pagan a sus empleados formalos sus quincenas.
- **Puente.** Identificación de los días viernes, sábado y domingo que anteceden a un día festivo que cae en lunes.

# Descripción general

A continuación se puede observar una descripción general de las variables más importantes de esta base de datos.

```{r Resumen de todas las variables}
print(dfSummary(accidentes), method = "render")
```

# Año

A continuación se observa una tabla de frecuencias para los años de los incidentes registrados en la base de datos de incidentes.

```{r Tabla de frecuencias para el año}
customGreen = "#c4ffff"

frecano <- as.data.frame(freq(accidentes$AÑO))

names(frecano)[1] <- "Frecuencia"
names(frecano)[2] <- "Válidos (%)"
names(frecano)[3] <- "Válidos acum. (%)"
names(frecano)[4] <- "Total acum. (%)"

tabano <- formattable(frecano, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
ano <- round(tabano, 2)
ano
```

Como se observa, los años 2015 a 2019 tienen aproximadamente la misma cantidad de observaciones en cada uno de ellos, lo cual es de esperar porque para todos ellos se cubre la totalidad del año, mientras que para los años 2014 y 2020 solo se cubre una parte de ellos. En todo caso, es interesante observar que el año en el que más accidentes se observaron fue el 2016, con 46,991 incidentes en total, lo cual representa el 17.38 % del total de observaciones. Gráficamente, se puede visualizar mediante el siguiente gráfico de barras:

```{r Gráfico de barras para el año}
g1 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = AÑO)) +
  ggtitle("Frecuencia de accidentes por año",
          subtitle = "Número de accidentes sucedidos por año en la ciudad de Medellín, Antioquia\nentre los años 2014 y 2020") +
  xlab("Año") +
  ylab("Frecuencia absoluta") +
  theme_minimal()
ggplotly(g1)
```

# Clase de accidente

La clase de accidente es una variable categórica de orden nominal que define de forma generar las características del incidente observado, de manera que esta variable puede asumir cualquiera de los siguientes niveles:

- Choque
- Atropello
- Volcamiento
- Caída del ocupante
- Incendio
- Otro

Teniendo esto presente, se puede definir la siguiente tabla de frecuencias para esta base de datos:

```{r Tabla de frecuencias para la clase de accidentes}
frecclase <- as.data.frame(freq(accidentes$CLASE_ACCIDENTE))

names(frecclase)[1] <- "Frecuencia"
names(frecclase)[2] <- "Válidos (%)"
names(frecclase)[3] <- "Válidos acum. (%)"
names(frecclase)[4] <- "Total acum. (%)"


tablaclase <- formattable(frecclase, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
clase <- round(tablaclase, 2)
clase
```

De la tabla anterior se puede observar que el tipo de accidente más frecuente es el choque, pues se observaron un total de 180,431 incidentes de tráfico que involucraron a este tipo de colisiones, lo que representa al 66.75 %  de los accidentes, lo cual es llamativo porque resulta más de dos terceras partes de todos los incidentes ocurridos en la ciudad de Medellín entre el año 2014 y 2020. A continuación se tienen a los incidentes que no clasifican en ninguna de las categorías propuestas por Informe Policial de Accidentes de Tránsito, IPAT, es decir, los que se etiquetan como "***otro***", dado que el 11,11 % de los incidentes registrados clasifican en esta categoría. En tercer lugar se tiene a los atropellamientos que abarcan al 9.36 % de los accidentes de tránsito observados en Medellín en el periodo mencionado. Por último, vale destacar que el tipo de incidente menos frecuente de todos es el que involucra a incendios, ya que durante este periodo solo sucedieron 35 eventos de este tipo, lo cual se evidencia en el siguiente gráfico de barras, pues la barra asociada a este tipo de incidente ni siquiera es apreciable, toda vez que las colisiones opacan al resto por su alta frecuencia relativa.

```{r Gráfico de barras para la clase de accidentes}
g2 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = CLASE_ACCIDENTE)) +
  ggtitle("Frecuencia de la clase de accidente",
          subtitle = "Frecuencia de los tipos de accidentes sucedidos en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Clase de accidente") +
  ylab("Frecuencia absoluta") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45))
ggplotly(g2)
```

# Diseño

Uno de los aspectos más importantes a la hora de evaluar un incidente de tránsito es caracterizar el tipo de espacio en el que sucedió el accidente en términos de la infraestructura física, y esta característica es determinada en la base de datos en la variable "***diseño***", que resulta siendo una variable categórica con un nivel nominal, y se describen los siguientes tipos de infraestructura:

- Ciclorruta.
- Glorieta.
- Intersección.
- Lote o predio.
- Paso a nivel.
- Paso elevado.
- Paso inferior.
- Pontón.
- Puente.
- Tramo de vía.
- Túnel.
- vía peatonal.

Y una tabla de frecuencias relacionada a esta variable es la siguiente:

```{r Tabla de frecuencias para el diseño}
frecdiseno <- as.data.frame(freq(accidentes$DISEÑO))

names(frecdiseno)[1] <- "Frecuencia"
names(frecdiseno)[2] <- "Válidos (%)"
names(frecdiseno)[3] <- "Válidos acum. (%)"
names(frecdiseno)[4] <- "Total acum. (%)"

tabdiseno <- formattable(frecdiseno, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
diseno <- round(tabdiseno, 2)
diseno
```

Revisando la tabla anterior se tiene que la infraestructura vial en la que más incidentes de tránsito sucedieron en Medellín entre el 2014 y el 2020 es el ***tramo de vía***, pues en estos se dieron el 69,45 % del total de accidentes observados en la ciudad de Medellín en el periodo mencionado, lo cual es bastante lejano al resto de incidentes. Además, es llamativo observar que el segundo tipo de infraestructura donde más incidentes se dieron son las intersecciones, donde ocurrieron el 16.30 % de los incidentes. Además, donde menos incidentes ocurrieron fue en los pasos inferiores, ya que únicamente el 0.25 % de los accidentes ocurrieron en este tipo de vías.

```{r Gráfico de barras para el diseño}
g3 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = DISEÑO)) +
  ggtitle("Tipo de infraestructa en la que se dio el incidente",
          subtitle = "Frecuencia del tipo de infraestructura en el que sucedieron en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Tipo de infraestructura") +
  ylab("Frecuencia absoluta") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45))
ggplotly(g3)
```

# Fecha

A continuación se muestra un gráfico en el que se muestra la cantidad de accidentes que se dieron en la ciudad de Medellín en cada uno de los días en los que se observaron datos. Para poder ver con mayor detalle algún periodo particular, se recomienda hacer un zoom con las herramientas que aparecen en el extremo superior derecho del gráfico.

```{r, echo = FALSE, message = FALSE}
# Creación de una base de datos que considera únicamente las fechas
accFechas <- accidentes %>% 
  group_by(FECHA) %>% 
  summarise(TOTAL = n())

fig <- plot_ly(accFechas, type = 'scatter', mode = 'lines')%>%
  add_trace(x = ~FECHA, y = ~TOTAL)%>%
  layout(showlegend = F)
fig <- fig %>%
  layout(
         xaxis = list(zerolinecolor = '#ffff',
                      zerolinewidth = 2,
                      gridcolor = 'ffff'),
         yaxis = list(zerolinecolor = '#ffff',
                      zerolinewidth = 2,
                      gridcolor = 'ffff'),
         plot_bgcolor='#e5ecf6', width = 900)


fig
```

Y del esquema anterior llama la atención que la mayoría de días se presentan entre 75y 175 accidentes por día, con algunas excepciones que se salen de estos extremos, de las cuales vale la pena mencionar que aquellas que aquellos picos superiores, es decir, fechas en las que se producen más accidentes en comparación con el resto de días, la mayoría se concentra en los meses de agosto y septiembre. Por otro lado, se debe rescatar que al comenzar cada año también se observa que ocurrencia de mínimos. También llama la atención la caída de accidentes que se presentan en la ciudad de Medellín luego del comienzo del confinamiento nacional como medida de contingencia por la pandemia de la COVID-19 el 18 de marzo de 2020.

Además, de estos datos se pueden rescatar los siguientes estadísticos, excluyendo los datos posteriores al 18 de marzo de 2020 por haber menos tránsito de vehículos que en otros días:

```{r, include = FALSE}
# Accidentalidad prepandemia
accFechasP <- accFechas %>% 
  filter(FECHA < as.Date("2020-03-18"))
kable(descr(accFechasP$TOTAL), digits = 2)
```

| Parámetro                | Valor |
|--------------------------|-------|
| Media                    | 115   |
| Desviación estándar      | 27.31 |
| Mínimo                   | 23    |
| Máximo (Q4)              | 99    |
| Primer cuantil (Q1)      | 99.5  |
| Mediana (Q2)             | 120   |
| Tercer cuantil           | 135.5 |
| Rango intercuartídico    | 35.25 |
| Coeficiente de variación | 0.24  |
| Coeficiente de asimetría | -0.44 |
| Curtosis                 | -0.09 |

Y como se observa, el número promedio de accidentes por día es 125 con una desviación estándar de 28.45 accidentes diarios. Además, el día en que menos accidentes ocurrieron fue el 4 de julio de 2014, donde se dieron 24 accidentes (nótese que este es justamente el día en el que se comienzan a tomar datos), toda vez que el día con mayor accidentes fue 4 de agosto de 2021, ya que en él se registraron 234 incidentes viales en la ciudad. Por otro lado, se tiene que la mediana es de 130 accidentes por día, lo que significa que en la mitad de los días se tienen 130 incidentes o más, mientras que en la otra mitad hay 129 accidentes de tránsito o menos. Asimismo, el primer cuantil ocurre en los 107.5 accidentes por día, y el tercer cuantil está asociado a los 144 accidentes viales diarios. Finalmente, se puede observar que el índice de asimetría es -0.4, lo que significa que los datos no son perfectamente simétricos y que tienden más hacia los valores más altos de accidentalidad diarios; además, la curtosis es de -0.06, lo cual significa que la cantidad de accidentes diarios podría tener una distribución parecida a la normal, y el hecho de que sea negativo implica que hay una concentración ligeralmente mayos datos hacia la media que en una distribución normal, lo que implica que la ocurrencia de valores extremos es menos frecuente que en una distribución auténticamente normal. Respecto a estos dos últimos parámetros, resulta interesante realizar un histograma para poder observar la distribución de accidentes diarios.

```{r}
barras <- 1 + 3.33*log(nrow(accFechas))
g7 <- ggplot(data = accFechas,
             aes(x = TOTAL)) +
        geom_histogram(aes(y = (..count..)/sum(..count..)),
                       fill = "darkblue",
                       bins = barras) +
        scale_y_continuous(labels=percent) +
        ggtitle("Histograma de la cantidad de accidentes diarios en Medellín") +
        xlab("Cantidad de accidentes diarios") +
        ylab("Frecuencia absoluta")
ggplotly(g7)
```

Y en este se observa un gráfico que no se asemeja al de una distribución normal, empezando por el hecho de que la simetría no es muy clara y también por la presencia de dos modas, lo cual no se asemeja a una distribución normal en el sentido que esta tiene una y solo una moda. En todo caso, vale la pena realizar una prueba de hipótesis de normalidad para esta variable mediante el test de Lillifors, que es el test más potente para una alta cantidad de datos y que se basa en una modificación del test de Lillifors. Para ello, se plantean las siguientes hipótesis con un nivel de confianza del 95 % ($\alpha = 0.05$):

$H_0:$ La distribución de la cantidad de accidentes por día en Medellín entre julio de 2014 y marzo de 2020 es normal.

$H_1:$ La distribución de la cantidad de accidentes por día en Medellín entre julio de 2014 y marzo de 2020 **no** es normal.

Y al realizarlo con ayuda de $\color{blue}{\textsf{R}}$ se obtiene un valor p de $V_p < 2.2 \times 10^{-16} < 0.05 = \alpha$, lo cual permite rechazar la hipótesis nula y por tanto se concluye la cantidad de accidentes que suceden en Medellín por día no tienen una distribución normal.

```{r, include = FALSE}
lillie.test(accFechas$TOTAL)
```

Ahora bien, para poder tener una mejor visibilidad, se va a hacer la suma de accidentes diarios pero presentados por mes.

```{r}
accFechasM <- accFechas %>% 
  mutate(MES = format(as.Date(FECHA), "%Y-%m")) %>% 
  group_by(MES) %>% 
  summarise(`Total mes` = sum(TOTAL))
```

```{r}
fig2 <- plot_ly(accFechasM, type = 'scatter', mode = 'lines')%>%
  add_trace(x = ~MES, y = ~`Total mes`)%>%
  layout(showlegend = F)
fig <- fig %>%
  layout(
         xaxis = list(zerolinecolor = '#ffff',
                      zerolinewidth = 2,
                      gridcolor = 'ffff'),
         yaxis = list(zerolinecolor = '#ffff',
                      zerolinewidth = 2,
                      gridcolor = 'ffff'),
         plot_bgcolor='#e5ecf6', width = 900)


fig2
```

Y en esta se puede constatar lo que ya se había mencionado previamente y es que existen mínimos en los meses de enero y máximos en los meses de agosto. Además, llama la atención que hay comportamiento aproximadamente uniforme en todos los años, salvo en el 2019, donde el pico del mes de agosto se preserva para los meses de septiembre y octubre. Además, se debe observar que en febrero del año 2020 se tiene un comportamiento semejante al de año previos hasta el mes de marzo, donde se da inicio al confinamiento nacional por cuenta de la pandemia del virus de la COVID-19. De aquí se pueden obtener los siguientes estadísticos excluyendo los valores posteriores a marzo de 2020:

```{r, include = FALSE}
accFechasMP <- accFechasM %>% 
  filter(MES < "2020-03")
kable(descr(accFechasMP$`Total mes`), digits = 2)
```

| Parámetro                | Valor    |
|--------------------------|----------|
| Media                    | 3,521.60 |
| Desviación estándar      | 262.63   |
| Mínimo                   | 2,920    |
| Máximo (Q4)              | 4,105    |
| Primer cuantil (Q1)      | 3,339.50 |
| Mediana (Q2)             | 3,531.5  |
| Tercer cuantil           | 3,694    |
| Rango intercuartídico    | 348      |
| Coeficiente de variación | 0.07     |
| Coeficiente de asimetría | -0.13    |
| Curtosis                 | -0.35    |

Y de lo anterior se sigue que la cantidad de accidentes promedio por mes es de 3,800, con una desviación estándar 281.27 accidentes mensuales. Además, se observa que el mínimo es de 3,142 accidentes mensuales, lo cual ocurre en junio de 2019; toda vez que el máximo es de 4,405 accidentes mensuales asociado a agosto de 2019. Además, la mediana es de 3,832 accidentes mensuales, lo que quiere decir que en al menos la mitad de los meses de los cuales hay registros cuentan con al menos 3,832 accidentes mensuales observados, toda vez que en la otra mitad se observaron una cantidad inferior de accidentes.

#### Mapa de calor

A continuación se observa un mapa de calor para la cantidad de accidentes observados por día.

```{r, out.width="150%", message = FALSE, echo = FALSE}
library("plyr")
accFechas <- na.omit(accFechas)
accFechas$weekday = as.POSIXlt(accFechas$FECHA)$wday #finding the day no. of the week
accFechas$weekdayf<-factor(accFechas$weekday,levels=rev(1:7),
                           labels=rev(c("L","M","X","J","V","S","D"))
                           ,ordered=TRUE) # converting the day no. to factor

accFechas$monthf<-factor(month(accFechas$FECHA),levels=as.character(1:12),
                         labels=c("Ene","Feb","Mar","Abr","May","Jun","Jul",
                                  "Ago","Sep","Oct","Nov","Dec"),
                         ordered=TRUE) # finding the month

accFechas$yearmonth<- factor(as.yearmon(accFechas$FECHA)) # finding the year and the month from the date. Eg: Nov 2018

accFechas$week <- as.numeric(format(accFechas$FECHA,"%W")) # finding the week of the year for each date

accFechas<-ddply(accFechas,.(yearmonth),transform,monthweek=1+week-min(week)) # normalizing the week to start at 1 for every month

p <- ggplot(accFechas, aes(monthweek, weekdayf, fill = accFechas$TOTAL)) + 
     geom_tile(colour = "white") + facet_grid(year(accFechas$FECHA)~monthf)+
      scale_fill_gradient(low="green", high="red") +
      xlab("Semana del mes") +
      ylab("") +
      ggtitle("Mapa de calor para los accidentes diarios") +
      labs(fill = "Acc. diarios")

p
detach("package:plyr", unload=TRUE)
```

En este resalta inicialmente quea partir del mes de marza de 2020 la cantidad de accidentes registrada por día en la última quincena de marzo y en el mes de abril es muy baja, toda vez que a medida que se va permitiendo la apertura de ciertos sectores económicos fue aumentado progresivamente la acidentalidad. Además, llama la atención que es en el mes de agosto donde se tienen más colores naranjas y oscuras, lo cual indica una alta cantidad de accidentailidad, mientras que en la primera quincena de enera predominan colores suaves y verdosos, que se asocian a bajas tasas de accidentalidad.

# Mes

Ahora se va a realizar la evaluación de la cantidad de accidentes viales ocurridos en la ciudad de Medellín según el mes sin importar el año, para lo cual se puede realizar la siguiente tabla de frecuencias:

```{r Tabla de frecuencias para el mes}
frecmes <- as.data.frame(freq(accidentes$MES))

names(frecmes)[1] <- "Frecuencia"
names(frecmes)[2] <- "Válidos (%)"
names(frecmes)[3] <- "Válidos acum. (%)"
names(frecmes)[4] <- "Total acum. (%)"

tabmes <- formattable(frecmes, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
mes <- round(tabmes, 2)
mes
```

La cual se puede acompañar del siguiente gráfico de barras.

```{r}
g6 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = fct_infreq(MES))) +
  ggtitle("Tipo de infraestructa en la que se dio el incidente",
          subtitle = "Frecuencia del tipo de infraestructura en el que sucedieron en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Tipo de infraestructura") +
  ylab("Frecuencia absoluta") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45))
ggplotly(g6)
```

Y como se observa, el mes en el que más accidentes se registran es agosto, que cuenta con 27,001 accidentes observados en dicho mes, seguido por 25,018 incidentes viales que ocurrieron durante los diferentes meses de julio, y en tercer lugar se tiene al mes de septiembre que registra a 24,129 accidentes. Además, los dos meses con menos incidentes viales registrados son enero con 20,098 incidentes viales y abril con 19,293 accidentes. Finalmente, es llamativo notar que los seis últimos meses del año se reparten los seis primeros lugares y que los meses del primer semestre hacen lo propio con los últimos seis puestos.    

# Comuna

A continución se visualiza la tabla de frecuencias para la cantidad de incidentes de tránsito registrados en las dieciséis comunas y cinco corregimientos de la ciudad de Medellín:

```{r Tabla de frecuencias para la comuna}
frecomuna <- as.data.frame(freq(accidentes$COMUNASMOD))

frecomuna <- frecomuna %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabcomuna <- formattable(frecomuna, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabcomuna, 2)
comuna
```

Y se acompaña con el siguiente gráfico de barras:

```{r}
g7 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = fct_infreq(COMUNASMOD))) +
  ggtitle("Tipo de infraestructa en la que se dio el incidente",
          subtitle = "Frecuencia del tipo de infraestructura en el que sucedieron en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Tipo de infraestructura") +
  ylab("Frecuencia absoluta") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45))
ggplotly(g7)
```

De los esquemas anteriores llama la atención que la comuna en la que más incidentes viales ocurren es en La Candelaria, que es la comuna central de la ciudad de Medellín y que concentra a la mayoría de las sedes de instituciones públicas municipales, departamentales nacionales, así como múltiples negocios y empresas que requieren una ingente cantidad de trabajadores y asimismo atrae a miles de consumidores diariamente. Así, se tiene que en dicha comuna se dan 67,352 incidentes viales o 24.91 % del total, lo que significa que casi uno de cada cuatro accidentes suceden en la comuna de La Candelaria. A continuación, se tiene a la comuna de Laureles-Estadio, la cual registró 27,978 incidentes viales en su territorio. En el otro extremo se tiene que los cuatro territorios en los que menos incidentes viales son corregimientos, en orden de mayor a menor: San Cristóbal (2,758 accidentes de tránsito), Santa Elena (487 accidentes de tránsito), Alta Vista (457 incidentes de tránsito) y San Sebastián de Palmitas (17 incidentes de tránsito), teniendo esta última 457 accidentes de tránsito. Adicionalmente, llama la atención que el corregimiento de San Antonio de Prado, que sumó 4,313 accidentes de tránsito, tuvo más incidentes que dos comunas: Popular (3,822 incidentes de tránsito) y Santa Cruz (3,437 accidentes de tránsito).

# Barrio

Respecto a los barrios, se tiene que los barrios en los que más accidentes de tránsito son Corazón de Jesús (La Candelaria) con 16,084 accidentes de tránsito, La Candelaria (La Candelaria) coon 5,915 accidentes de tránsito y Caribe (Castilla) con  5,619 incidentes de tránsito en total, siendo el último caracterizado por estar en las inmediaciones de la Terminal de Transportes Norte de la ciudad de Medellín, así como varios intercambios hacia los cuatro puntos cardinales de la ciudad, así como salidas hacia la autopista Medellín - Bogotá y vías que van hacia el Caribe colombiano y hacia el occidente antioqueño.

```{r Tabla de frecuencias para el barrio}
barriosAux <- accidentes %>% # Vector con todos los barrios de la BD
  select(c(BARRIO))
barrios <- barriosAux %>% 
  group_by(BARRIO) %>% 
  summarise(Frecuencia = n())

datatable(barrios,
          colnames = c("Número de entrada",
                       "Nombre del barrio",
                       "Frecuencia absoluta"))
```

# Variables adicionales

En el capítulo de transformaciones adicionales a la base de dato se explicaron algunas variables explicativas adicionales a partir de la fecha para ser empleadas en la creación del modelo predictivo que permita proyectar los accidentes para los años 2020 y 2021. En ese sentido, se va a mostrar básico de lo que se obtuvo en cada una de estas variables extra. Debe tenerse en cuenta que como algunas de las variables creadas con dicotómicas binarias, es decir, que sus niveles de medición son únicamente el número uno o el número cero, entonces todos los resultados que se verán a continuación están en términos de proporciones o porcentajes.

## Día modificado

A continuación se observa una tabla en la que se evidencia la cantidad total y la proporción de accidentes que ocurren en cada uno de los días de un mes.

```{r Tabla de frecuencias para el día modificado}
frecdiax <- as.data.frame(freq(accidentes$DIAX))

frecdiax <- frecdiax %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabdiax <- formattable(frecdiax, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabdiax, 2)
comuna
```

Y como se  observa, hay una distribución aproximadamente uniforme en la cantidad de accidentes que sucede por día, ya que la proporción de incidentes viales que suceden en cada día del mes aproximadamente la misma para cada día. Esto se puede ver mejor mediante el siguiente gráfico de barras.

```{r}
g8 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = fct_infreq(as.factor(DIAX)))) +
  ggtitle("Día del mes en que se dio el incidente de tránsito",
          subtitle = "Frecuencia del día del mes en el que sucedieron en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Día del mes") +
  ylab("Frecuencia absoluta") +
  theme_minimal()
ggplotly(g8)
```

En este gráfico son notables las caídas en la cantidad de incidentes viables que se dan en los últimos días del mes, es decir, los días 29, 30 y 31. Sin embargo, esto puede ser atribuible al hecho de que el mes de febrero llega hasta el 28 de febrero (salvo en los años 2016 y 2020), lo que cual puede explicar la reducción en los días 29 y 30, y respecto al día 31 se tiene que tener presente que solo siete de los doce meses del mes tienen 31 días, de manera que se omiten las observaciones para el resto de los cinco meses y a partir de ello se puede explicar que su proporción sea aproximadamente la mitad del resto de día.

## Semana modificada (natural)

Entiéndase por semana natural aquella que siempre tendrá como primera semana del año aquella que abarca los primeros siete días del año, es decir, del primero a siete de enero, independientemente de los días de la semana en los que hayan ocurrido dichos días. En el siguiente gráfico se observa la cantidad de incidentes viables acontencidos en cada semana del año.

```{r}
g9 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = fct_infreq(as.factor(SEMANAX)))) +
  ggtitle("Número de la semana en que se dio el incidente de tránsito",
          subtitle = "Frecuencia del número de la semana del año en el que sucedieron en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Semana del año") +
  ylab("Frecuencia absoluta") +
  theme_minimal() +
  theme(axis.text.x = element_text(angle = 45))
ggplotly(g9)
```

De este gráfico llama especialmente la atención que la semana en la que más incidentes viales se dieron, que es la número 31 (29 de julio al 4 de agosto), destacaba notablemente del resto, pero se tiene que tener presente que las otras semanas que completan las cinco semanas con más incidentes del año son aquellas que se durante la última quincena de julio y el mes de agosto, lo cual se podría explicar por el hecho de que durante es mes ocurre la Feria de Flores de Medellín, se paga la primera prima legal del año a los empleados formales de la ciudad y múltiples cadenas con importante presencia en la ciudad (como los almacenes Éxito) celebran su aniversario con descuentos. Por otro lado, si se omite la semana 53 del año (que únicamente incluye a los días 30 y 31 de diciembre), la semana con menos incidentes viales del año es la primera, lo cual se puede atribuir a la disminución del tráfico vehícular en la ciudad de Medellín dadas las vacaciones de millones de empleados y estudiantes de toda la ciudad.

## Día de la semana

```{r Tabla de frecuencias para el día modificado2}
frecdiaxx <- as.data.frame(freq(accidentes$DIA_SEMANA))

frecdiaxx <- frecdiaxx %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabdiaxx <- formattable(frecdiaxx, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabdiaxx, 2)
comuna
```

De la tabla anterior se puede destacar que hay una distribución aproximadamente uniforme para los incidentes que ocurren en cada uno de los días de la semana, con un 15 % de incidentes por día, salvo el día domingo que registra el 9.35 % de los incidentes viales; empero, de los otros seis días, destaca el día viernes con un 15.92 % de los incidentes registrados, lo cual podría deberse a la gran cantidad de personas que se transportan por la ciudad no solo para atender actividades académicas, laborales y comerciales, entre otras, durante la mañana y la tarde, sino también por la salida de personas de la ciudad este día o por la búsqueda de actividades de ocio en la noche. La tabla anterior se puede ilustrar mejor a través de la siguiente gráfica de barras:

```{r}
g10 <-ggplot(data = accidentes) +
        geom_bar(mapping = aes(x = fct_infreq(as.factor(DIA_SEMANA)))) +
  ggtitle("Día de la semana en que se dio el incidente de tránsito",
          subtitle = "Frecuencia del día de la semana en la que sucedieron en los incidentes de tránsito\nen la ciudad de Medellín, Antioquia entre los años 2014 y 2020") +
  xlab("Día del mes") +
  ylab("Frecuencia absoluta") +
  theme_minimal()
ggplotly(g10)
```

## Día festivo

```{r Tabla de frecuencias para el día festivo}
frecfest <- as.data.frame(freq(accidentes$FESTIVO))

frecfest <- frecfest %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabfest <- formattable(frecfest, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabfest, 2)
comuna
```

Se debe comenzar teniendo presente que aquellos días que coinciden con un día festivo son aquellos que están marcados con el número uno (1), mientras que los que no tienen el número cero (0). En virtud de lo anterior, se observa que el 2.84 % de los incidentes viales observados en Medellín entre agosto de 2014 y marzo de 2021 ocurrieron un día festivo.

## Feria de Flores

```{r Tabla de frecuencias para el día de feria de flores}
frecfestf <- as.data.frame(freq(accidentes$FLORES))

frecfestf <- frecfestf %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabfestf <- formattable(frecfestf, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabfestf, 2)
comuna
```

Se debe comenzar teniendo presente que aquellos días que coinciden con un día en el que se celebró la feria de flores en Medellín son aquellos que están marcados con el número uno (1), mientras que los que no tienen el número cero (0). En virtud de lo anterior, se observa que el 2.6 % de los incidentes viales observados en Medellín entre agosto de 2014 y marzo de 2021 ocurrieron durante la feria de flores.

## Feria de Flores

```{r Tabla de frecuencias para la quincena}
festquinc <- as.data.frame(freq(accidentes$QUINCENA))

festquinc <- festquinc %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabquinc <- formattable(festquinc, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabquinc, 2)
comuna
```

Se debe comenzar teniendo presente que aquellos días que coinciden con un día en el que se pagó quincena (días 1, 5, 15, 20 o 30 de cada mes) son aquellos que están marcados con el número uno (1), mientras que los que no tienen el número cero (0). En virtud de lo anterior, se observa que el 15,97 % de los incidentes viales observados en Medellín entre agosto de 2014 y marzo de 2021 ocurrieron durante la cancelación de la quincena de los empleados formales de la ciudad.


## Días previos a un festivo

```{r Tabla de frecuencias para la prefestivo}
festprev <- as.data.frame(freq(accidentes$PREFESTIVO))

festprev <- festprev %>% 
  rename("Frecuencia" = "Freq") %>% 
  rename("Válidos (%)" = "% Valid") %>% 
  rename("Válidos acum. (%)" = "% Valid Cum.") %>% 
  rename("Total (%)" = "% Total") %>% 
  rename("Total acum. (%)" = "% Total Cum.")

tabprev <- formattable(festprev, list(
  `Frecuencia` = color_bar(customGreen),
  `Válidos (%)` = color_bar(customGreen),
  `Válidos acum. (%)` = color_bar(customGreen),
  `Total (%)` = color_bar(customGreen),
  `Total acum. (%)` = color_bar(customGreen)))
comuna <- round(tabprev, 2)
comuna
```

Se debe comenzar teniendo presente que aquellos días que son viernes, sábado y domingo y que ocurren previo a un lunes festivo son aquellos que están marcados con el número uno (1), mientras que los que no tienen el número cero (0). En virtud de lo anterior, se observa que el 8.07 % de los incidentes viales observados en Medellín entre agosto de 2014 y marzo de 2021 ocurrieron durante un viernes, sábado y domingo y que ocurren previo a un lunes festivo.
