==> BD_nueva/6_Agrupamiento.Rmd <==
---
title: "Agrupamiento de los barrios de Medellín según el nivel de accidentalidad"
author: "Equipo de TAE"
date: "Diciembre de 2021"
output:
  html_document:
    code_folding: show
    toc: true
    theme: paper
    df_print: paged
    number_sections: true
---

<style>
body {
text-align: justify}
</style>

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

# Agrupando Barrios de Medellín por accidentes


```{r, include = FALSE}
library("tidyverse")
library("summarytools")
library("DT")
```


```{r}
load("accidentesMDE2.RData")
```

```{r}
barrios <- accidentes %>% 
  group_by(BARRIO) %>% 
  summarise(CANTIDAD = n())
```


```{r}
cantidades <- as.vector(barrios$CANTIDAD)
```

Para la preparación del agrupamiento de los barrios por accidente, se separó una base de datos aparte donde se tenían las variables a estudiar, tales como barrio, gravedad de accidente y la cantidad de accidentes por barrio. Se realizó un histograma de la variable cantidades para separar por rangos de alto número de accidentes, moderado y bajo, y así, con la media y la desviación estándar realizamos esta división.

```{r}
datatable(descr(cantidades))
hist(cantidades,
     main = "Histograma del número de accidentes por barrio")
```

```{r}
media = 793
sd = 979
max = 5917
k1 = round((max - media)/sd)
cortes <- as.numeric(character())

for (i in 1:k1) {
    cortes <- c(cortes, media + i * sd)
}
cortesx <- c("Muy bajo", "Bajo", "Moderado", "Alto", "Muy alto", "Grave")
```

Luego, para tener en cuenta la gravedad de accidentes por barrio se creó un índice de gravedad puesto que esta estaba como una variable categórica, este índice se creó de la siguiente manera: sea $x$ el número de fallecidos que hay por incidentes viales en un barrio, $y$ la cantidad de accidentes viales con heridos en el mismo barrio y $z$ la suma de incidentes que solo dejaron daños materiales. Entonces, el índice de gravedad se define así:

$$índiceGravedad(Barrio) = \frac{0.7x + 0.25*y + 0.05 z}{x + y+ z}$$

```{r, include = FALSE}
agrupador <- accidentes %>%
        group_by(BARRIO) %>%
          summarise(TOTAL = n(),
                    muertos = length(GRAVEDAD_ACCIDENTE[GRAVEDAD_ACCIDENTE == "Con muertos"]),
                    heridos = length(GRAVEDAD_ACCIDENTE[GRAVEDAD_ACCIDENTE == "Con heridos"]),
                    danos = length(GRAVEDAD_ACCIDENTE[GRAVEDAD_ACCIDENTE == "Solo daños"]),
                    indiceGravedad = (0.7 * muertos + 0.25 * heridos + 0.05 * danos)/TOTAL)
descr(agrupador$indiceGravedad)
```

```{r, echo = FALSE, warning = FALSE}
ggplot(data = agrupador, mapping = aes(y = indiceGravedad, x = TOTAL)) +
  geom_point()
```

```{r, echo = FALSE, warning = FALSE}
auxiliar <- agrupador %>% 
  select(c(BARRIO, TOTAL, indiceGravedad))
head(auxiliar)
```

```{r, echo = FALSE, warning = FALSE}
mT <- mean(auxiliar$TOTAL)
sdT <- sd(auxiliar$TOTAL)
miG <- mean(auxiliar$indiceGravedad)
sdiG <- sd(auxiliar$indiceGravedad)
auxiliar <- auxiliar %>% 
  mutate(TOTALc = (TOTAL - mT)/sdT) %>% 
  mutate(iGc = (indiceGravedad - miG)/sdiG)
```

```{r, include = FALSE, warning = FALSE}
library("factoextra")
```

```{r, echo = FALSE, warning = FALSE}
grupos <- fviz_nbclust(auxiliar[, c(4, 5)], kmeans, method = "wss")
grupos
k = 3
```

```{r, echo = FALSE, warning = FALSE}
library("stats")
```

```{r, echo = FALSE, warning = FALSE}
k3 <- kmeans(auxiliar[, c(4, 5)], k)
clusteres <- as.data.frame(k3[1])
auxiliar <- cbind(auxiliar, clusteres)
auxiliar <- auxiliar %>% 
  mutate(cluster = recode(cluster,
                          "1" = "Alto",
                          "2" = "Moderado",
                          "3" = "Bajo"))
```


```{r}
f <- ggplot(data = auxiliar, mapping = aes(y = indiceGravedad, x = TOTAL)) +
  geom_point(aes(color = cluster))
f
```

Teniendo estas dos variables numéricas se busca el K óptimo para la efectiva división de la cantidad de grupos que debe haber para hacer el mapa de color, también se creó un vector extra para agregar a la base de datos en orden de cómo deben aparecer los barrios en el mapa y su respectiva agrupación (alto, moderado, bajo y no aplica para aquellos en los barrios en los que no se registró accidente). A partir de la gráfica anterior, entonces se puede definir que los barrios con índice de gravedad *bajo* (en verde) son aquellos que tienen bajas cantidades de incidentes y que estos no suelen ser muy graves, mientras que los que tienen índice de gravedad *alto* se destacan por tener una gran cantidad de incidentes viales pero no necesariamente muy graves, siendo los de categoría *moderada* aquellos que presentan poco incidentes viales pero con saldos fatales con frecuencia.

```{r, echo = FALSE, warning = FALSE}
auxiliar <- auxiliar %>% 
  select(c(BARRIO, cluster))

names(auxiliar[2]) <- "Riesgo"
```

```{r, echo = FALSE, warning = FALSE}
save(auxiliar, file = "agrupamiento.Rmd")
write.csv(auxiliar,"agrupamiento.csv")
```

```{r, echo = FALSE, warning = FALSE}
rm(list = ls())
```

```{r, include = FALSE, warning = FALSE}
library("leaflet")
library("rgdal")
library("mapview")
```

```{r, echo = FALSE, warning = FALSE}
barriosMDE <- readOGR(".",layer="Limite_Barrio_Vereda_Catastral")
aux <- as.data.frame(barriosMDE)
```

```{r}
# Cambio de nombres
barriosMDE$NOMBRE_BAR[5] <- "Los Balsos No. 1"
barriosMDE$NOMBRE_BAR[9] <- "El Nogal - Los Almendros"
barriosMDE$NOMBRE_BAR[15] <- "La Mota"
barriosMDE$NOMBRE_BAR[19] <- "Playón de los Comuneros"
barriosMDE$NOMBRE_BAR[20] <- "El Velódromo"
barriosMDE$NOMBRE_BAR[29] <- "Área de expansión San Antonio de Prado"
barriosMDE$NOMBRE_BAR[32] <- "Belalcázar"
barriosMDE$NOMBRE_BAR[33] <- "Los Balsos No. 2"
barriosMDE$NOMBRE_BAR[35] <- "Suburbano Travesías"
barriosMDE$NOMBRE_BAR[48] <- "Bomboná No. 1"
barriosMDE$NOMBRE_BAR[52] <- "Alfonso López"
barriosMDE$NOMBRE_BAR[55] <- "San Martín de Porres"
barriosMDE$NOMBRE_BAR[56] <- "Santa Lucía"
barriosMDE$NOMBRE_BAR[61] <- "Andalucía"
barriosMDE$NOMBRE_BAR[63] <- "Fátima"
barriosMDE$NOMBRE_BAR[67] <- "Campo Valdés No. 2"
barriosMDE$NOMBRE_BAR[70] <- "Los Ángeles"
barriosMDE$NOMBRE_BAR[76] <- "Doce de Octubre No. 2"
barriosMDE$NOMBRE_BAR[83] <- "Alejandro Echavarría"
barriosMDE$NOMBRE_BAR[87] <- "La Mansión"
barriosMDE$NOMBRE_BAR[92] <- "Bermejal - Los Álamos"
barriosMDE$NOMBRE_BAR[95] <- "El Corazón - El Morro"
barriosMDE$NOMBRE_BAR[96] <- "Berlín"
barriosMDE$NOMBRE_BAR[99] <- "Bomboná No. 2"
barriosMDE$NOMBRE_BAR[109] <- "San José de la Cima No. 1"
barriosMDE$NOMBRE_BAR[112] <- "La América"
barriosMDE$NOMBRE_BAR[125] <- "El Corazón"
barriosMDE$NOMBRE_BAR[130] <- "Las Lomas No. 2"
barriosMDE$NOMBRE_BAR[132] <- "San José de la Cima No. 2"
barriosMDE$NOMBRE_BAR[133] <- "Barrio de Jesús"
barriosMDE$NOMBRE_BAR[135] <- "Santa María de los Ángeles"
barriosMDE$NOMBRE_BAR[138] <- "Diego Echavarría"
barriosMDE$NOMBRE_BAR[141] <- "Área de expansión Altavista"
barriosMDE$NOMBRE_BAR[142] <- "Aguas Frías"
barriosMDE$NOMBRE_BAR[144] <- "San Javier No. 1"
barriosMDE$NOMBRE_BAR[148] <- "Cabecera Urbana San Cristóbal"
barriosMDE$NOMBRE_BAR[151] <- "San Germán"
barriosMDE$NOMBRE_BAR[156] <- "Nueva Villa de Aburrá"
barriosMDE$NOMBRE_BAR[166] <- "San Javier No. 2"
barriosMDE$NOMBRE_BAR[167] <- "Área de expansión Belén Rincón"
barriosMDE$NOMBRE_BAR[170] <- "Área de expansión Altos de Calasanz"
barriosMDE$NOMBRE_BAR[176] <- "Suburbano La Palma - El Patio"
barriosMDE$NOMBRE_BAR[178] <- "Estación Villa"
barriosMDE$NOMBRE_BAR[185] <- "Campo Valdés No. 2"
barriosMDE$NOMBRE_BAR[187] <- "Travesías"
barriosMDE$NOMBRE_BAR[192] <- "López de Mesa"
barriosMDE$NOMBRE_BAR[194] <- "Suburbano Aguas Frías"
barriosMDE$NOMBRE_BAR[195] <- "María Cano Carambolas"
barriosMDE$NOMBRE_BAR[210] <- "El Progreso No. 2"
barriosMDE$NOMBRE_BAR[215] <- "Unidad Deportiva Atanasio Girardot"
barriosMDE$NOMBRE_BAR[219] <- "El Rincón"
barriosMDE$NOMBRE_BAR[220] <- "Barrio Cerro El Volador"
barriosMDE$NOMBRE_BAR[225] <- "Santa Inés"
barriosMDE$NOMBRE_BAR[230] <- "Loma de los Bernal"
barriosMDE$NOMBRE_BAR[233] <- "San Joaquín"
barriosMDE$NOMBRE_BAR[234] <- "Suburbano Pedregal Alto"
barriosMDE$NOMBRE_BAR[237] <- "Cataluña"
barriosMDE$NOMBRE_BAR[238] <- "Jardín Botánico"
barriosMDE$NOMBRE_BAR[249] <- "Simón Bolívar"
barriosMDE$NOMBRE_BAR[251] <- "San José del Manzanillo"
barriosMDE$NOMBRE_BAR[257] <- "La Piñuela"
barriosMDE$NOMBRE_BAR[261] <- "Moscú No. 1"
barriosMDE$NOMBRE_BAR[264] <- "Suburbano El Plan"
barriosMDE$NOMBRE_BAR[267] <- "Suburbano El Tesoro"
barriosMDE$NOMBRE_BAR[270] <- "Hospital San Vicente de Paúl"
barriosMDE$NOMBRE_BAR[271] <- "Nueva Villa de la Iguaná"
barriosMDE$NOMBRE_BAR[273] <- "Belén"
barriosMDE$NOMBRE_BAR[276] <- "Alejandría"
barriosMDE$NOMBRE_BAR[278] <- "Los Alcázares"
barriosMDE$NOMBRE_BAR[280] <- "San José de la Montaña"
barriosMDE$NOMBRE_BAR[286] <- "Boyacá"
barriosMDE$NOMBRE_BAR[287] <- "Córdoba"
barriosMDE$NOMBRE_BAR[289] <- "Santa Mónica"
barriosMDE$NOMBRE_BAR[293] <- "Corazón de Jesús"
barriosMDE$NOMBRE_BAR[307] <- "Barrio Colón"
barriosMDE$NOMBRE_BAR[314] <- "La Palma - El Patio"
barriosMDE$NOMBRE_BAR[315] <- "Área de expansión Pajarito"
barriosMDE$NOMBRE_BAR[321] <- "Santa Fe"
barriosMDE$NOMBRE_BAR[327] <- "Doce de Octubre No. 2"
barriosMDE$NOMBRE_BAR[329] <- "Moscú No. 2"
barriosMDE$NOMBRE_BAR[331] <- "Jesús Nazareno"
barriosMDE$NOMBRE_BAR[339] <- "Batallón Girardot"
barriosMDE$NOMBRE_BAR[346] <- "Héctor Abad Gómez"
barriosMDE$NOMBRE_BAR[348] <- "Antonio Nariño"
```

```{r, include = FALSE, warning = FALSE}
m=leaflet(barriosMDE)
m=addTiles(m)
```

```{r, include = FALSE, warning = FALSE}
m=addPolygons(m,popup=barriosMDE$NOMBRE_BAR)
m=addTiles(m)
```

```{r, echo = FALSE, warning = FALSE}
vector <- c("Bajo","Moderado","Moderado","Moderado","Bajo",
            "Alto","No aplica","Bajo","Bajo","Moderado",
            "Bajo","Bajo","Moderado","Moderado","Bajo",
            "Moderado","Bajo","Moderado","Moderado","Bajo",
            "Bajo","Moderado","No aplica","Moderado","Alto",
            "Bajo","Bajo","Bajo","Moderado","Moderado",
            "Bajo","Bajo","Bajo","Moderado","Moderado",
            "Moderado","Alto","Moderado","Bajo","Moderado",
            "No aplica","Bajo","Alto","Bajo","Moderado",
            "Moderado","Alto","Bajo","Moderado","Bajo",
            "Bajo","Moderado","Bajo","Moderado","Moderado",
            "Moderado","Moderado","Moderado","Moderado","bajo",
            "Moderado","Bajo","Bajo","Moderado","Moderado",
            "Bajo","Moderado","Moderado","Moderado","Bajo",
            "Moderado","Alto","Moderado","Moderado","Moderado",
            "Moderado","Moderado","No aplica","Alto","Bajo",
            "Alto","Alto","Moderado","Bajo","Moderado",
            "Moderado","Moderado","Bajo","Alto","Moderado",
            "No aplica","Moderado","Bajo","Moderado","Moderado",
            "Moderado","Moderado","Moderado","Moderado","Moderado",
            "Bajo","Moderado","Moderado","Moderado","Moderado",
            "Moderado","Moderado","Bajo","Moderado","Moderado",
            "Bajo","Bajo","Bajo","No aplica","Bajo",
            "Moderado","Bajo","Bajo","Alto","Alto",
            "Bajo","Bajo","Moderado","Alto","Moderado",
            "Moderado","Alto","Moderado","Moderado","Bajo",
            "Bajo","Moderado","Moderado","Bajo","Bajo",
            "Moderado","Moderado","Bajo","No aplica","No aplica",
            "Moderado","Moderado","Moderado","Moderado","Bajo",
            "Moderado","No aplica","Moderado","Moderado","Bajo",
            "Bajo","Moderado","Alto","Bajo","Moderado",
            "Bajo","Bajo","Moderado","Moderado","Moderado",
            "Moderado","Bajo","Moderado","Bajo","Bajo",
            "Moderado","Bajo","Bajo","Moderado","Bajo",
            "Moderado","Moderado","Moderado","Bajo","Bajo",
            "Bajo","No aplica","Bajo","Moderado","Moderado",
            "Alto","Moderado","Moderado","Bajo","Moderado",
            "Moderado","Bajo","Bajo","Alto","Moderado",
            "Alto","Moderado","Bajo","Moderado","Moderado",
            "No aplica","Moderado","Bajo","Moderado","Moderado",
            "Moderado","Moderado","Moderado","Bajo","Bajo",
            "Bajo","Alto","Alto","Moderado","Moderado",
            "No aplica","Bajo","Moderado","Moderado","Bajo",
            "Bajo","Bajo","Bajo","Bajo","Moderado",
            "Moderado","Bajo","Moderado","Moderado","Moderado",
            "Moderado","Moderado","Moderado","Bajo","Bajo",
            "Bajo","Bajo","Bajo","Moderado","Moderado",
            "Bajo","Bajo","Bajo","Moderado","No aplica",
            "Moderado","Bajo","No aplica","Moderado","No aplica",
            "Moderado","Bajo","Moderado","Bajo","Moderado",
            "Bajo","Alto","Moderado","Moderado","Bajo",
            "Moderado","Moderado","No aplica","No aplica","No aplica",
            "Moderado","Alto","Bajo","Moderado","Moderado",
            "Moderado","Bajo","No aplica","Bajo","Bajo",
            "Bajo","Bajo","Alto","Moderado","Moderado",
            "Bajo","Moderado","Moderado","Moderado","Moderado",
            "Bajo","Moderado","Moderado","Moderado","Moderado",
            "Moderado","Moderado","Bajo","Moderado","Moderado",
            "Moderado","Moderado","Alto","Bajo","Bajo",
            "Moderado","Moderado","Bajo","Moderado","Moderado",
            "Bajo","Bajo","Moderado","Bajo","Alto",
            "Alto","Alto","Alto","Moderado","Moderado",
            "Moderado","No aplica","Alto","No aplica","Moderado",
            "No aplica","Moderado","Bajo","Moderado","Moderado",
            "Alto","Moderado","Moderado","Moderado","Bajo",
            "Alto","Moderado","Bajo","Moderado","Moderado",
            "Alto","Moderado","Moderado","Moderado","Moderado",
            "No aplica","Moderado","No aplica","Moderado","Moderado",
            "Bajo","Bajo","No aplica","Moderado","Moderado",
            "Bajo","Moderado","Moderado")
```

```{r, echo = FALSE, warning = FALSE}
aux["Riesgo"] <- vector
```

```{r, echo = FALSE, warning = FALSE}
#d90000 ALTO
#ffa333 MODERADO
#006400 BAJO
#808080 NA
aux["colores"] <- rep(NA, size = nrow(aux))
aux$colores[aux$Riesgo == "Alto"] <- "#d90000"
aux$colores[aux$Riesgo == "Moderado"] <- "#ffa333"
aux$colores[aux$Riesgo == "Bajo"] <- "#006400"
aux$colores[aux$Riesgo == "No aplica"] <- "#808080"
```

Con lo anterior se crea el siguiente mapa, donde se muestra los barrios de Medellín agrupados por colores donde se especifica cómo se divide la ciudad respecto a los accidentes que ocurren en Medellín, con esto se podría tener una idea de en qué lugares de la ciudad es más probable un accidente de alta gravedad, moderada o baja.

```{r, echo = FALSE, warning = FALSE}
#colores=sample(x=c("blue","green","yellow", "red", "pink"),size=length(barriosMDE),replace=TRUE)
m = addPolygons(m,
                color = aux$colores,
                fillOpacity = 0.7,
                popup = paste0(
                "<b>Barrio: </b>",
                barriosMDE$NOMBRE_BAR,
                "<br>",
                 "<b>Comuna: </b>",
                barriosMDE$NOMBRE_COM,
                "<br>",
                "<b>Nivel de riesgo: </b>",
                aux$Riesgo))
m = addLegend(m, "bottomright",
    colors = c("#d90000",  "#ffa333", "#006400", "#808080"),
    labels = c("Alto", "Moderado","Bajo","No aplica"),
    title = "Nivel de riesgo",
    opacity = 1)
m
```
